I/O Interrupts dev

+++ src/esvm/vm/instructions/Out.java
@@ -5,10 +5,10 @@ package esvm.vm.instructions;
  */
 public class Out extends Instruction{
     public byte arg1;
-    public short arg2;
+    public int arg2;
     public final static byte code = 29;
 
-    public Out(byte arg1, short arg2) {
+    public Out(byte arg1, int arg2) {
         this.arg1 = arg1;
         this.arg2 = arg2;
         asm = "Out";
